requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pandas==2.1.3
numpy==1.26.2
faiss-cpu==1.7.4
google-generativeai==0.3.1
python-dotenv==1.0.0
scikit-learn==1.3.2
requests==2.31.0
streamlit==1.28.2